Reorganize some unit test files
