+++ tests/IntegrationTests/EpdContacts/UpdateStatusTests.cs
@@ -1,0 +1,62 @@
+﻿using EnfoTests.Infrastructure.Helpers;
+using EnfoTests.TestData;
+using FluentAssertions;
+using NUnit.Framework;
+using System;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace EnfoTests.Infrastructure.EpdContacts;
+
+public class UpdateStatusTests
+{
+    [Test]
+    [TestCase(true)]
+    [TestCase(false)]
+    public async Task UpdateStatus_IfChangeStatus_Succeeds(bool newActiveStatus)
+    {
+        var itemId = EpdContactData.EpdContacts.First(e => e.Active != newActiveStatus).Id;
+
+        using var repositoryHelper = RepositoryHelper.CreateRepositoryHelper();
+        using var repository = repositoryHelper.GetEpdContactRepository();
+
+        await repository.UpdateStatusAsync(itemId, newActiveStatus);
+        repositoryHelper.ClearChangeTracker();
+
+        var item = await repository.GetAsync(itemId);
+        item.Active.Should().Be(newActiveStatus);
+    }
+
+    [Test]
+    [TestCase(true)]
+    [TestCase(false)]
+    public async Task UpdateStatus_IfStatusUnchanged_Succeeds(bool newActiveStatus)
+    {
+        var itemId = EpdContactData.EpdContacts.First(e => e.Active == newActiveStatus).Id;
+
+        using var repositoryHelper = RepositoryHelper.CreateRepositoryHelper();
+        using var repository = repositoryHelper.GetEpdContactRepository();
+
+        await repository.UpdateStatusAsync(itemId, newActiveStatus);
+        repositoryHelper.ClearChangeTracker();
+
+        var item = await repository.GetAsync(itemId);
+        item.Active.Should().Be(newActiveStatus);
+    }
+
+    [Test]
+    public async Task UpdateStatus_FromMissingId_ThrowsException()
+    {
+        const int itemId = -1;
+
+        var action = async () =>
+        {
+            using var repository = RepositoryHelper.CreateRepositoryHelper().GetEpdContactRepository();
+            await repository.UpdateStatusAsync(itemId, true);
+        };
+
+        (await action.Should().ThrowAsync<ArgumentException>())
+            .WithMessage($"ID ({itemId}) not found. (Parameter 'id')")
+            .And.ParamName.Should().Be("id");
+    }
+}
