+++ tests/IntegrationTests/LegalAuthorities/UpdateTests.cs
@@ -1,0 +1,78 @@
+﻿using Enfo.Domain.LegalAuthorities.Resources;
+using EnfoTests.Infrastructure.Helpers;
+using EnfoTests.TestData;
+using FluentAssertions;
+using NUnit.Framework;
+using System;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace EnfoTests.Infrastructure.LegalAuthorities;
+
+public class UpdateTests
+{
+    [Test]
+    public async Task Update_FromValidItem_Updates()
+    {
+        var itemId = LegalAuthorityData.LegalAuthorities.First(e => e.Active).Id;
+        var resource = new LegalAuthorityCommand { Id = itemId, AuthorityName = "New Name" };
+
+        using var repositoryHelper = RepositoryHelper.CreateRepositoryHelper();
+        using var repository = repositoryHelper.GetLegalAuthorityRepository();
+
+        await repository.UpdateAsync(resource);
+        repositoryHelper.ClearChangeTracker();
+
+        var item = await repository.GetAsync(itemId);
+        item.Should().BeEquivalentTo(resource);
+    }
+
+    [Test]
+    public async Task Update_WithNoChanges_Succeeds()
+    {
+        var original = LegalAuthorityData.LegalAuthorities.First(e => e.Active);
+        var resource = new LegalAuthorityCommand { Id = original.Id, AuthorityName = original.AuthorityName };
+
+        using var repositoryHelper = RepositoryHelper.CreateRepositoryHelper();
+        using var repository = repositoryHelper.GetLegalAuthorityRepository();
+
+        await repository.UpdateAsync(resource);
+        repositoryHelper.ClearChangeTracker();
+
+        var item = await repository.GetAsync(original.Id);
+        item.Should().BeEquivalentTo(new LegalAuthorityView(original));
+    }
+
+    [Test]
+    public async Task Update_FromInvalidItem_ThrowsException()
+    {
+        var itemId = LegalAuthorityData.LegalAuthorities.First(e => e.Active).Id;
+        var resource = new LegalAuthorityCommand { Id = itemId, AuthorityName = null };
+
+        var action = async () =>
+        {
+            using var repository = RepositoryHelper.CreateRepositoryHelper().GetLegalAuthorityRepository();
+            await repository.UpdateAsync(resource);
+        };
+
+        (await action.Should().ThrowAsync<ArgumentException>())
+            .And.ParamName.Should().Be(nameof(LegalAuthorityCommand.AuthorityName));
+    }
+
+    [Test]
+    public async Task Update_WithMissingId_ThrowsException()
+    {
+        const int itemId = -1;
+        var itemUpdate = new LegalAuthorityCommand { Id = itemId, AuthorityName = "New Name" };
+
+        var action = async () =>
+        {
+            using var repository = RepositoryHelper.CreateRepositoryHelper().GetLegalAuthorityRepository();
+            await repository.UpdateAsync(itemUpdate);
+        };
+
+        (await action.Should().ThrowAsync<ArgumentException>())
+            .WithMessage($"ID ({itemId}) not found. (Parameter 'resource')")
+            .And.ParamName.Should().Be("resource");
+    }
+}
